% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAQC_checks.R
\name{clock_QAQC}
\alias{clock_QAQC}
\title{Checks time of computer clock by extracting true time from internet}
\usage{
clock_QAQC(input)
}
\arguments{
\item{input}{time-sync data generated from function}
}
\description{
Checks time of computer clock by extracting true time from internet
}
\details{
identifies and colorcodes time differnce between computer and official internet time
}
