% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_schema.R
\name{col_class_out}
\alias{col_class_out}
\title{Schema that defines column data type output for "read_vdat_csv" function}
\usage{
col_class_out(in_col_name)
}
\arguments{
\item{in_col_name}{vector of column names}
}
\value{
vector of data types that can be passed on to 'colClasses' argument in data.table::fread
}
\description{
This function looks up appropriate data types for ONLY columns needd and used in QAQC app.  Many columns output in vdat are not included(!).  Data.table::fread will guess the appropriate data types for columns not included in this table.
}
