% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_vdats.R
\name{compile_vdats}
\alias{compile_vdats}
\title{Converts vrl to csv using Innovasea vdat command line utility.}
\usage{
compile_vdats(vdat_files, v_path, temp_dir = tempdir())
}
\arguments{
\item{vdat_files}{a character vector of one or more file paths pointing to the vrl files that will be included in the download report}

\item{v_path}{a character file path pointing to the vdat utility.}

\item{temp_dir}{A folder where converted vdat files will be saved.  Defaults to system temporary directory.}
}
\value{
A list object that contains one element for each vrl submitted in \code{vdat_files}
}
\description{
Converts vrl to csv using Innovasea vdat command line utility.
}
\examples{
# load example vrl
vrl <- system.file("extdata/VR2W_109412_20190619_1.vrl", package = "glatosQAQC")

# set system path to vdat utility 
if(Sys.info()["sysname"] == "Linux") vpath <- system.file("exec/vdat_linux", package = "glatosQAQC")

# compile vrl files

tst <- vrl_to_csv(vrl_file = vrl_file, out_dir = tempdir(), vdat_path = vpath)


foo <- compile_vdats(vdat_files = vrl, v_path = vpath)
vdat_files = c(
"C:/Users/thayden/Desktop/QAQC_weirdness/VR2AR_547562_20220906_1.vrl",
"C:/Users/thayden/Desktop/QAQC_weirdness/VR2Tx_480029_20220906_1.vrl",
"C:/Users/thayden/Desktop/QAQC_weirdness/VR2Tx_483620_20220906_1.vrl",
"C:/Users/thayden/Desktop/QAQC_weirdness/VR2Tx_483626_20220906_1.vrl",
"C:/Users/thayden/Desktop/QAQC_weirdness/VR2Tx_483638_20220906_1.vrl",
"C:/Users/thayden/Desktop/QAQC_weirdness/VR2Tx_483813_20220906_1.vrl" ,
"C:/Users/thayden/Documents/VR2AR_546310_20190607_1.vrl")
vdat_files <- system.file("extdata/VR2W_109412_20190619_1.vrl", package = "glatosQAQC")

vdat_files <- c(
"C:/Users/thayden/Documents/VR2AR_Samples_with_Gen2/new/VR2AR_546906_20230516_1.vrl",
"C:/Users/thayden/Documents/VR2AR_Samples_with_Gen2/new/VR2AR_547544_20230516_1.vrl"
)

vdat_files <- c(
"C:/Users/thayden/Documents/VR2AR_Samples_with_Gen2/new/VR2AR_546906_20230516_1.vrl"
)

v_path = c("C:/Program Files/Innovasea/Fathom/vdat.exe")
temp_dir = tempdir()
foo <- compile_vdats(vdat_files = vdat_files, v_path = v_path, temp_dir = temp_dir)
}
\author{
Todd Hayden
}
