% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAQC_checks.R
\name{excel_QAQC}
\alias{excel_QAQC}
\title{Colorcodes potential errors in excel output}
\usage{
excel_QAQC(input, output)
}
\arguments{
\item{input}{output from "process table", runs checks and makes excel file}

\item{output}{file path for excel worksheet}
}
\description{
Colorcodes potential errors in excel output
}
\details{
uses R package openxlsx to identify potential errors in vrl files and colorizes them using custom formatting in excel
}
