% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_records.R
\name{extract_records}
\alias{extract_records}
\title{Extract records contained in VRL files for report}
\usage{
extract_records(vdat, type = "DET")
}
\arguments{
\item{vdat}{List of vdat files from csv created by \code{compile_vdats}}

\item{type}{name of data types to be extracted.}
}
\value{
a data.table containing all of records extracted from csv file
}
\description{
Extract records contained in VRL files for report
}
\examples{


# load example vrl
vrl <- system.file("extdata/VR2W_109412_20190619_1.vrl", package = "glatosQAQC")

# set system path to vdat utility 
if(Sys.info()["sysname"] == "Linux") vpath <- system.file("vdat/vdat_linux", package = "glatosQAQC")

# compile vrl files
df <- compile_vdats(vrl, vpath)

# extract detections
df <- extract_records(df, type = "DET")


}
\author{
Todd Hayden
}
