% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_table.R
\name{fls}
\alias{fls}
\title{Compiles output data from all receiver log files, creates table used in output report}
\usage{
fls
}
\arguments{
\item{fls}{data.frame of produced from shiny::fileInput that contains name, size, type, datapath for each receiver selected by user}

\item{action}{User selected radio button for "download" and "initialize" included in output table}

\item{batt}{User selected radio button for logging status of battery at time of report}

\item{vdat_pth}{Path to vdat executable installed on your computer}
}
\value{
returns table of diagnostic  metrics extracted from each receiver
}
\description{
Compiles output data from all receiver log files, creates table used in output report
}
