% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatosQAQC-package.R
\docType{package}
\name{glatosQAQC-package}
\alias{glatosQAQC}
\alias{glatosQAQC-package}
\title{glatosQAQC}
\description{
This package uses vdat executable to extract summary information from Innovasea acoustic receiver log files using an interactive app.
File summary and summary figures are downloadable, validity checks are applied to identify issues.
}
\author{
\strong{Maintainer}: Todd Hayden \email{thayden@usgs.gov}

Authors:
\itemize{
  \item Todd Hayden \email{thayden@usgs.gov}
}

}
\keyword{internal}
