% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_detections.R
\name{process_detections}
\alias{process_detections}
\title{utility function that extracts and summarizes receiver download data needed to produce download report##'}
\usage{
process_detections(input)
}
\arguments{
\item{input}{data.table produced by \code{extract_records}}
}
\value{
processed data frame
}
\description{
utility function that extracts and summarizes receiver download data needed to produce download report##'
}
\author{
Todd Hayden
}
