% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_table.R
\name{process_table}
\alias{process_table}
\title{Compiles data from all receivers, generates table of data included in report.}
\usage{
process_table(
  fls,
  mrk_params,
  nme,
  action,
  work_dir = work_dir,
  datapath = datapath
)
}
\arguments{
\item{fls}{a character vector of files to include in report}

\item{work_dir}{working directory where processing takes place
fls <- c("/home/thayden/Desktop/VR2AR_546908_20230814_1.vrl", "/home/thayden/Desktop/VR2AR_547553_20230814_1.vrl")

mrk_params = "/home/thayden/tools/vdat"
work_dir = "C:/Users/thayden/Desktop"
nme <- c("VR2AR_546908_20230814_1.vrl", "VR2AR_547553_20230814_1.vrl")
action <- "down"
datapath = file.path(tempdir(), c("0.vrl", "1.vrl"))
dtc <- glatosQAQC::compile_vdats(vdat_files = fls, v_path = pth, temp_dir = "C:/Users/thayden/Desktop")
foo <-  process_table(fls = fls, mrk_params = mrk_params, nme = nme, action = action, datapath = datapath, work_dir = "/home/thayden/Desktop")}

\item{params}{rmarkdown parameters needed to process data}
}
\description{
Compiles data from all receivers, generates table of data included in report.
}
