% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAQC_checks.R
\name{QAQC}
\alias{QAQC}
\title{Complete check of extracted detection data}
\usage{
QAQC(out)
}
\arguments{
\item{out}{extracted data from VRL as created by running a process_table function}
}
\description{
Complete check of extracted detection data
}
