% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{QAQCapp}
\alias{QAQCapp}
\title{Start the glatosQAQC app.}
\usage{
QAQCapp(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{runApp}}}
}
\value{
Interactive application object app.
}
\description{
This function starts the interactive web app built by this package.
}
\details{
This is a basic function which does only one thing: starts the interactive app.  There are no input arguments.
}
\examples{
if(interactive()){
QAQCapp()
}
}
