% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-read_vdat_csv.r
\name{read_vdat_csv}
\alias{read_vdat_csv}
\title{Read detections and events from an Innovasea Fathom VDAT CSV file}
\usage{
read_vdat_csv(src, record_types = NULL)
}
\arguments{
\item{src}{A character string with path and name of a Innovasea VDAT CSV
detection file. If only file name is given, then the file must be located
in the working directory.}

\item{record_types}{An optional vector of character strings with names of
record types to read from the file. E.g., "DET" for detection records.
Default (\code{NULL}) will read all record types present in input CSV
\code{src}.}
}
\value{
A list of class \code{vdat_list} with one named element for each
record type and attributes: \code{fathom_csv_format_version} with version
of the input Fathom CSV format; \code{vat_exe_version} = with version of
VDAT.exe used to create input CSV.
}
\description{
Read detections and events from a Innovasea Fathom VDAT CSV file
}
\details{
Reading is done via \code{\link[data.table]{fread}}.

All timestamp columns are assumed to be in UTC.
}
\examples{
vrl_file <- system.file("extdata",
  "VR2W_109924_20110718_1.vrl", package="glatos")

csv_file <- vrl_to_csv(vrl_file, out_dir = dirname(vrl_file))

#read all record types
vdat <- read_vdat_csv(csv_file)

#read only two record types
vdat <- read_vdat_csv(csv_file, record_types = c("DATA_SOURCE_FILE", "DET"))

}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
