% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_compare.R
\name{time_compare}
\alias{time_compare}
\title{time comparison function- pulls time information from internet (if available) and calculates difference between computer (system) time and internet. If these are different, then computer was not synchronized prior to download.}
\usage{
time_compare()
}
\description{
time comparison function- pulls time information from internet (if available) and calculates difference between computer (system) time and internet. If these are different, then computer was not synchronized prior to download.
}
