% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{vdat_i_tag}
\alias{vdat_i_tag}
\title{constructor function for integrated tag}
\usage{
vdat_i_tag(x)
}
\description{
constructor function for integrated tag
}
\examples{
i_tag <- glatosQAQC::extract_records(vdat = dtc, type = "CFG_TRANSMITTER")
vdat_i_tag(i_tag)
}
