% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-vrl_to_csv.r
\name{vrl_to_csv}
\alias{vrl_to_csv}
\title{Convert a Vemco VRL file to a Vemco Fathom CSV file}
\usage{
vrl_to_csv(
  vrl_file = NULL,
  out_dir = NULL,
  vdat_path = "C:/Program Files (x86)/VEMCO/Fathom/vdat.exe"
)
}
\arguments{
\item{vrl_file}{An optional character string with path and name of a
Vemco VRL detection file, or a vector of VRL file names. If only file name
is given, then the file must be located in the working directory.}

\item{out_dir}{A character string directory where CSV files will be written.
If \code{NULL} (default) then file(s) will be written to the working
directory (e.g., \code{getwd()}).}

\item{vdat_path}{The full path to \code{VDAT} command line program.
The must path includes file executable file (i.e., */.vdat.exe for Windows)}

\item{vrl_dir}{An optional character string with path to a directory
containing one or more Vemco VRL detection files.}

\item{overwrite}{Logical. If TRUE, output CSV file(s) will overwrite existing
CSV file(s) with same name in \code{out_dir}. When FALSE (default),'_n'
(i.e., _1, _2, etc.) will be appended to names of output files that already
exist in \code{outDir}. \strong{NOT YET IMPLEMENTED.}}

\item{recursive}{Logical. Indicates if VRL files in subdirectories
of \code{vrl_dir} should be converted. Default is \code{FALSE}.}

\item{show_progress}{Logical. Indicates if progress bar should be shown.}

\item{diag}{LOgical. Indicates if diagnostic output should be included
output (default = FALSE). If TRUE, a data.frame with file size, message
and status flag (from \code{system2} call).}
}
\value{
A character string or vector with the name(s) (including path) of
the CSV files created.
}
\description{
Use Vemco's VDAT command line program VDAT (distributed with Vemco Fathom
software) to make a CSV file containing detection data from a
Vemco VRL file in VDAT format.
}
\details{
Conversion is done by system call to the Vemco program VDAT.exe. You must
have VDAT.exe available at the location specified by \code{vdat_path}.
VDAT.exe is included with Vemco's Fathom software.
}
\section{Outputs}{

A text file in Vemco Fathom CSV format for each input VRL file.
Each csv is named the same (except for extension) as the source VRL file.
}

\examples{
#get path to example Vemco VRL file
vrl_file <- system.file("extdata", 
  "VR2W_109924_20110718_1.vrl", package="glatosDS")
vrl_to_csv(vrl_file, out_dir = dirname(vrl_file))

vrl_file <- c(
"C:/Users/thayden/Documents/VR2AR_Samples_with_Gen2/new/VR2AR_546906_20230516_1.vrl",
"C:/Users/thayden/Documents/VR2AR_Samples_with_Gen2/new/VR2AR_547544_20230516_1.vrl"
)
vdat_path = c("C:/Program Files/Innovasea/Fathom/vdat.exe")
out_dir = tempdir()
vrl_to_csv(vrl_file, out_dir = out_dir, vdat_path = vdat_path)
}
\author{
C. Holbrook (cholbrook@usgs.gov)
}
