% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-glatosQAQC.R
\docType{package}
\name{glatosQAQC}
\alias{glatosQAQC-package}
\alias{glatosQAQC}
\title{glatosQAQC: glatosQAQC shiny app to interactively check file validity of log files downloaded from Innovasea acoustic receivers}
\description{
Shiny app that extracts and summarizes data from Innovasea *.vrl or *.vdat files and presents a number of validity checks.
}
\author{
\strong{Maintainer}: Todd Hayden \email{thayden@usgs.gov}

}
