% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-glatosQAQC.R
\docType{package}
\name{glatosQAQC}
\alias{glatosQAQC}
\title{An R package for the QA/QC of vrl files downloaded from receivers}
\description{
\code{glatosQAQC} is an R package that interactively checks
validity quality of vrl files downloaded from Vemco acoustic receivers.  Package
produces a report of receiver statisitics that enables researcher
to check for errors. Vemco VRL files are read using VDAT command line utility.
}
\section{Package status}{

This package is in early development.
If you encounter problems or have questions or suggestions, please post an
issue or email thayden@usgs.gov (maintainer: Todd Hayden).
}

\section{Installation}{

Package is hosted on GitHub and can be installed automatically via \code{install_github} function in the \code{devtools} package.
If you do not have \code{devtools} installed on your machine, first run \code{install.packages("devtools")} in a R terminal  After package \code{devtools} is installed, then run \code{devtools::install_github("haydento/glatosQAQC")} to install \code{glatosQAQC}.

\emph{A recent version of Fathom Connect software (Innovasea) must be installed for \code{glatosQAQC} to work.  Fathom Connect can be downloaded at \url{https://support.vemco.com/s/downloads}}
}

\section{Typical Use}{

\cr
\emph{\code{glatosQAQC} consists of an interactive GUI that enables the user to open and extract summary data from one or more .vrl files stored on your computer.}
To use run the following code in a R terminal:
\enumerate{
\item \code{library(glatosQAQC)}\cr
\item \code{QAQCapp()}
\cr
\cr
This will open a separate window in your default browser that allows selection of .vrl files.  Once files are selected, a summary table will be generated.
}
}

