% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-glatosQAQC.R
\docType{package}
\name{glatosQAQC}
\alias{glatosQAQC}
\title{An R package for the QA/QC of vrl files downloaded from receivers}
\description{
\code{glatosQAQC} is an R package that interactively checks
validity quality of vrl files downloaded from Vemco acoustic receivers.  Package
produces a report of receiver statisitics that enables researcher
to check for errors. Vemco VRL files are read using VDAT command line utility.
}
\section{Package status}{

This package is in early development.
If you encounter problems or have questions or suggestions, please post an
issue or email thayden@usgs.gov (maintainer: Todd Hayden).
}

\section{Installation}{

To install:\cr\cr
TBD
}

\section{Typical Use}{

\cr
\enumerate{
\item \code{library(glatosQAQC)}\cr
\item \code{QAQCapp()}\cr\cr
}
}

