% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_detections.R
\name{process_detections}
\alias{process_detections}
\title{utility function that extracts and summarizes receiver download data needed to produce download report}
\usage{
process_detections(input, battery = FALSE)
}
\arguments{
\item{input}{data.table produced by \code{extract_records}}

\item{battery}{boolean, set battery = TRUE when processing records for battery voltage}
}
\value{
processed data frame
}
\description{
utility function that extracts and summarizes receiver download data needed to produce download report
}
\examples{

# load example vrl
vrl <- system.file("extdata/VR2W_109412_20190619_1.vrl", package = "glatosQAQC")

# set system path to vdat utility 
if(Sys.info()["sysname"] == "Linux") vpath <- system.file("exec/vdat_linux", package = "glatosQAQC")

# compile vrl files
df <- compile_vdats(vrl, vpath)

# extract detections
df <- extract_records(df, type = "DET")

# process detections necessary for download report generation.
out <- process_detections(df)


}
\author{
Todd Hayden
}
