% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_table.R
\name{process_table}
\alias{process_table}
\title{Compiles data from all receivers, generates table of data included in report.}
\usage{
process_table(
  fls,
  mrk_params,
  nme,
  action,
  work_dir = work_dir,
  datapath = datapath
)
}
\arguments{
\item{fls}{a character vector of files to include in report}

\item{work_dir}{working directory where processing takes place
fls <- c("c:/Users/thayden/Documents/VR2AR_Samples_with_Gen2/new/VR2AR_546906_20230516_1.vrl", "c:/Users/thayden/Documents/VR2AR_Samples_with_Gen2/new/VR2AR_547544_20230516_1.vrl")

mrk_params = "C:/Program Files/Innovasea/Fathom/vdat.exe"
work_dir = "C:/Users/thayden/Desktop"
nme <- c("VR2AR_546906_20230516_1.vrl", "VR2AR_547544_20230516_1.vrl")
action <- "down"
datapath = file.path(tempdir(), c("0.vrl", "1.vrl"))
foo <-  process_table(fls = fls, mrk_params = mrk_params, nme = nme, action = action, datapath = datapath, work_dir = "C:/Users/thayden/Desktop")}

\item{params}{rmarkdown parameters needed to process data}
}
\description{
Compiles data from all receivers, generates table of data included in report.
}
