% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_table.R
\name{process_table}
\alias{process_table}
\title{Compiles data from all receivers, generates table of data included in report.}
\usage{
process_table(
  fls,
  mrk_params,
  nme,
  action,
  work_dir = work_dir,
  datapath = datapath
)
}
\arguments{
\item{fls}{a character vector of files to include in report}

\item{work_dir}{working directory where processing takes place
fls <- c("~/Documents/VRL tests/VR2AR_546310_20190607_1.vrl", "~/Documents/VRL tests/VR2AR_546310_20220624_1.vrl", "~/Documents/VRL tests/VR2W_134214_20230626_1.vrl")
mrk_params = "C:\\PROGRA~1\\vdat\\vdat.exe"
work_dir = "~/Documents/"
nme <- c("VR2AR_546310_20190607_1.vrl", "VR2AR_546310_20220624_1.vrl", "VR2W_134214_20230626_1.vrl")
action <- "down"
datapath = file.path(tempdir(), c("0.vrl", "1.vrl"))
dtc <- glatosQAQC::compile_vdats(vdat_files = fls, v_path = mrk_params)

foo <-  process_table(fls = fls, mrk_params = mrk_params, nme = nme, action = action, datapath = datapath, work_dir = tempdir())}

\item{params}{rmarkdown parameters needed to process data}
}
\description{
Compiles data from all receivers, generates table of data included in report.
}
