% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_table.R
\name{process_table}
\alias{process_table}
\title{Compiles data from all receivers, generates table of data included in report.}
\usage{
process_table(
  fls,
  mrk_params,
  nme,
  action,
  work_dir = work_dir,
  datapath = datapath
)
}
\arguments{
\item{fls}{a character vector of files to include in report}

\item{work_dir}{working directory where processing takes place
fls <- c("C:/Users/thayden/Desktop/QAQC_weirdness/VR2AR_547562_20220906_1.vrl", "C:/Users/thayden/Desktop/QAQC_weirdness/VR2Tx_480029_20220906_1.vrl", "C:/Users/thayden/Documents/big_test/VR2W_106323_20140701_1.vrl")

fls <- c("C:/Users/thayden/Documents/big_test/VR2W_106324_20110727_1.vrl", "C:/Users/thayden/Documents/big_test/VR2W_106324_20140616_1.vrl", "C:/Users/thayden/Documents/big_test/VR2W_106323_20140701_1.vrl", "C:/Users/thayden/Desktop/QAQC_weirdness/VR2AR_547562_20220906_1.vrl")
mrk_params = "C:/Program Files/Innovasea/Fathom/vdat.exe"
work_dir = "C:/Users/thayden/Desktop"
nme <- c("VR2AR_547562_20220906_1.vrl", "VR2Tx_480029_20220906_1.vrl", "VR2W_106323_20140701_1.vrl")
action <- "down"
datapath = file.path(tempdir(), c("0.vrl", "1.vrl", "2.vrl"))}

\item{params}{rmarkdown parameters needed to process data}
}
\description{
Compiles data from all receivers, generates table of data included in report.
}
