% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_table.R
\name{process_table}
\alias{process_table}
\title{Compiles data from all receivers, generates table of data included in report.}
\usage{
process_table(fls, mrk_params, nme, action, work_dir = work_dir)
}
\arguments{
\item{fls}{a character vector of files to include in report}

\item{work_dir}{working directory where processing takes place}

\item{params}{rmarkdown parameters needed to process data}
}
\description{
Compiles data from all receivers, generates table of data included in report.
}
